Move general functions out of concantenation.
